Change Arkane default to not create plots
Previously, Arkane had the option to create plots, but if you didn't
add that option, it still creates plots. Thus there is no way to
not create the plots in the code, and the entire flag is pointless.
This causes issues since the plotting function can error and crash
an entire job (and there is no way to disable plotting)

This commit changes the default to not create plots, which it seems
like is how the code was designed to function. If the user/developer
wants plots, then they can use the argument '-p' to obtain plots.

See commit 1a0c2dec25b26a680828a7df7fc11dbd20ebc19b for previous info.