Arkane now uses Species.label for chemkin files.
This commit changes the default Arkane option to use the labels
supplied in the Arkane input file since an RMG run is not run
so there are not identifying indexes.

A warning is printed in both the log file and in the chemkin
output when overly long species labels and reactions occur.
This is not a perfect fix, but eliminates majority of errors
with Arkane chemkin files, has clarity, and will eventually
be replaced by yaml outputs.